refactor: clean up current-time observer
